Project folder structure change
