+++ src/functions/core/DataFunction.cpp
@@ -1,0 +1,2 @@
+#include "DataFunction.hpp"
+
