+++ src/functions/core/DataFunction.hpp
@@ -1,0 +1,10 @@
+#pragma once
+#include "Function.hpp"
+
+namespace MATD{
+  namespace FUNC{
+    class DataFunction : public Function {
+
+    };
+  }
+}
