+++ src/functions/core/Function.cpp
@@ -1,0 +1,1 @@
+#include "Function.hpp"
