+++ src/functions/core/ShaderFunction.cpp
@@ -1,0 +1,1 @@
+#include "ShaderFunction.hpp"
