+++ src/graphs/core/Node.cpp
