+++ src/graphs/data_graph/DataGraph.cpp
@@ -1,6 +1,6 @@
 #include "DataGraph.hpp"
-#include "../core/Core.hpp"
-#include "MaterialGraph.hpp"
+#include "../../core/Core.hpp"
+#include "../core/MaterialGraph.hpp"
 
 MATD::GRAPH::DataGraph::DataGraph(MaterialGraph* graph, const MATD::JSON& JSONObj) : MATD::GRAPH::Graph(graph, JSONObj)
 {
