+++ src/graphs/data_graph/DataGraph.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "Graph.hpp"
+#include "../core/Graph.hpp"
 
 namespace MATD{
   namespace GRAPH{
