+++ src/graphs/kernel_graph/KernelGraph.cpp
@@ -1,5 +1,5 @@
 #include "KernelGraph.hpp"
-#include "MaterialGraph.hpp"
+#include "../core/MaterialGraph.hpp"
 
 MATD::GRAPH::KernelGraph::KernelGraph(MATD::GRAPH::MaterialGraph* graph,const MATD::JSON& JSONObj) : MATD::GRAPH::Graph(graph, JSONObj)
 {
